School/requirements.txt
django==2.0
djangorestframework==3.8.2
psycopg2==2.7.4
django-flat-responsive==2.0
django-admin-interface
